Move JSON stuff to io package
