+++ pdtable/io/__init__.py
@@ -1,2 +1,4 @@
 from ._csv import read_csv, write_csv
 from ._excel import read_excel, write_excel
+from ._json import to_json_serializable, JsonData, JsonDataPrecursor
+from .json_converters import table_to_json_data, json_data_to_table
